+++ Runtime/Systems/InteractionSystem/Interactable.cs
@@ -5,7 +5,6 @@ using UnityEngine.Events;
 
 public class Interactable : MonoBehaviour
 {
-    public UnityEvent onInteractVoid;
     public UnityEvent<Interactor> onInteract;
 
     private List<Interactor> interactors;